Allow applications to use the "code id_token" response_type
This response type is required to build a FAPI compliant client, and as
the application can verify the id_token signature and c_hash before
calling the token endpoint there is no reason to block applications from
using code id_token.

Other response types are still blocked as they don't appear to make sense
in the context of a native app.

closes #292